move reducers back into apps
