+++ webview/src/plots/components/App.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import {
   PlotsData,
   PlotsDataKeys,
@@ -6,14 +5,33 @@ import {
   SectionCollapsed
 } from 'dvc/src/plots/webview/contract'
 import { MessageToWebview } from 'dvc/src/webview/contract'
+import React from 'react'
 import { Plots } from './Plots'
-import { setCollapsed as setCustomPlotsCollapsed } from './customPlots/customPlotsSlice'
-import { setCollapsed as setComparisonTableCollapsed } from './comparisonTable/comparisonTableSlice'
-import { setCollapsed as setTemplatePlotsCollapsed } from './templatePlots/templatePlotsSlice'
-import { initialize } from './webviewSlice'
+import {
+  setCollapsed as setComparisonTableCollapsed,
+  update as updateComparisonTable,
+  updateShouldShowTooManyPlotsMessage as updateShouldShowTooManyImagesMessage
+} from './comparisonTable/comparisonTableSlice'
+import {
+  setCollapsed as setCustomPlotsCollapsed,
+  update as updateCustomPlots
+} from './customPlots/customPlotsSlice'
+import {
+  setCollapsed as setTemplatePlotsCollapsed,
+  updateShouldShowTooManyPlotsMessage as updateShouldShowTooManyTemplatesMessage,
+  update as updateTemplatePlots
+} from './templatePlots/templatePlotsSlice'
+import {
+  initialize,
+  updateCliError,
+  updateHasPlots,
+  updateHasUnselectedPlots,
+  updatePlotErrors,
+  updateSelectedRevisions
+} from './webviewSlice'
 import { PlotsDispatch } from '../store'
 import { useVsCodeMessaging } from '../../shared/hooks/useVsCodeMessaging'
-import { dispatchAction } from '../../shared/dispatchAction'
+import { dispatchActions } from '../../shared/dispatchActions'
 
 const dispatchCollapsedSections = (
   sections: SectionCollapsed,
@@ -27,6 +45,23 @@ const dispatchCollapsedSections = (
     dispatch(setTemplatePlotsCollapsed(sections[PlotsSection.TEMPLATE_PLOTS]))
   }
 }
+
+const actionToDispatch = {
+  [PlotsDataKeys.CLI_ERROR]: updateCliError,
+  [PlotsDataKeys.CUSTOM]: updateCustomPlots,
+  [PlotsDataKeys.COMPARISON]: updateComparisonTable,
+  [PlotsDataKeys.TEMPLATE]: updateTemplatePlots,
+  [PlotsDataKeys.HAS_PLOTS]: updateHasPlots,
+  [PlotsDataKeys.HAS_UNSELECTED_PLOTS]: updateHasUnselectedPlots,
+  [PlotsDataKeys.PLOT_ERRORS]: updatePlotErrors,
+  [PlotsDataKeys.SELECTED_REVISIONS]: updateSelectedRevisions,
+  [PlotsDataKeys.SHOW_TOO_MANY_TEMPLATE_PLOTS]:
+    updateShouldShowTooManyTemplatesMessage,
+  [PlotsDataKeys.SHOW_TOO_MANY_COMPARISON_IMAGES]:
+    updateShouldShowTooManyImagesMessage
+} as const
+
+export type PlotsActions = typeof actionToDispatch
 
 export const feedStore = (
   data: MessageToWebview<PlotsData>,
@@ -45,7 +80,7 @@ export const feedStore = (
       dispatch
     )
   }
-  dispatchAction('plots', stateUpdate, dispatch)
+  dispatchActions(actionToDispatch, stateUpdate, dispatch)
 }
 
 export const App = () => {
